+++ tests/__init__.py
@@ -4,6 +4,10 @@ metaschema_uri_2019_09 = URI("https://json-schema.org/draft/2019-09/schema")
 metaschema_uri_2020_12 = URI("https://json-schema.org/draft/2020-12/schema")
 metaschema_uri_next = URI("https://json-schema.org/draft/next/schema")
 
+core_vocab_uri_2019_09 = URI("https://json-schema.org/draft/2019-09/vocab/core")
+core_vocab_uri_2020_12 = URI("https://json-schema.org/draft/2020-12/vocab/core")
+core_vocab_uri_next = URI("https://json-schema.org/draft/next/vocab/core")
+
 example_schema = {
     "$schema": "https://json-schema.org/draft/2020-12/schema",
     "$id": "dynamicRef8_main.json",